Moving the simulators to a unique folder for this them.
