+++ spicelib/simulators/ngspice_simulator.py
@@ -23,7 +23,7 @@ from pathlib import Path
 import logging
 _logger = logging.getLogger("spicelib.NGSpiceSimulator")
 
-from .simulator import Simulator, run_function
+from spicelib.sim.simulator import Simulator, run_function
 
 
 class NGspiceSimulator(Simulator):
